daos

+++ app/src/main/java/nsu/krpo/academads/data/network/models/BookingRequest.kt
@@ -1,13 +1,14 @@
 package nsu.krpo.academads.data.network.models
 
 import com.squareup.moshi.Json
+import java.math.BigDecimal
 
 data class BookingRequest(
         @Json(name = "ads_id")
         val adsId: Long,
 
         @Json(name = "claimant")
-        val claimant: Long,
+        val claimant: BigDecimal,
 
         @Json(name = "dateUntil")
         val dateUntil: String
